feat: add dialect#normaliseidentifier and dialect#parserelationref (#120)
# Description

- **Introducing `Dialect#NormaliseIdentifier`:** normalises the
identifier's parts that are unquoted, typically by lower or upper casing
them, depending on the database.
- **Introducing `Dialect#ParseRelationRef `:** parses a string into a
`RelationRef` after normalising the identifier and stripping out
surrounding quotes. The result is a `RelationRef` with case-sensitive
fields, i.e. it can be safely re-quoted in full (see [QuoteTable] and,
for instance, used for matching against the database's information
schema.
- **Deprecating `Dialect#FormatTableName`** due to wrong normalisation
behaviour in `bigquery` and `mysql`.

## Security

- [x] The code changed/added as part of this pull request won't create
any security issues with how the software is being used.