chore(redshift): retry schema does not exist errors in list schemas (#208)
# Description

If the list schemas query is executed while a schema is being deleted,
the query can fail with a schema does not exist error. This is happening
for instance while multiple pull request tests are running. Therefore,
we are now retrying such errors.

## Security

- [x] The code changed/added as part of this pull request won't create
any security issues with how the software is being used.